Batch rendering for trees, one huge water layer (not water per chunk)
Me and GPT fixed the batch rendering for trees. Now its a few calls (1 or 2) instead of thousands to the GPU. I went from 20 FPS to more than 60 FPS because of this. I also added a big water layer instead of each chunk having its own water layer. This removed the dark seem where chunks meet. Note that we need around 100+ vertices for water in order for the fog to interpolate accurately. I also added a Enemy model which tracks the player. But this is still very basic